micro: Handle +/text search text from args
This is a feature found in vim and commonly used
by Linux kernel test robots to give context about
warnings and/or failures.

e.g. vim +/imem_size +623 drivers/net/ipa/ipa_mem.c

Implement a similar behavior that searches for the (last)
match of +/$text in the opened buffers.